+++ test/libsolidity/smtCheckerTests/functions/payable_1.sol
@@ -1,0 +1,15 @@
+contract C {
+	function g() external {
+		f();
+	}
+
+	function h() external payable {
+		f();
+	}
+
+	function f() internal {
+		require(msg.value == 0);
+	}
+}
+// ====
+// SMTEngine: all